python-backend/ENTERPRISE_PROJECT/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-jose
pydantic>=2.0
pydantic-settings
celery
redis
pytest
httpx
